logistic-header created and working with ShopperPage (orders)
