+++ src/pages/shopper/shopper.html
@@ -6,18 +6,20 @@
 -->
 
 <ion-header no-border>
-  <logistic-header></logistic-header>
+  <logistic-header
+    (ordersToPage)="displayOrders($event)">
+    </logistic-header>
 </ion-header>
 
 <ion-content>
 
   <!--Actual list-->
 
-  <h1 [hidden]="availableOrders.length > 0">Pas de commandes {{openShippings?'ouvertes':'fermées'}} pour ce mois-ci</h1>
+  <h1 [hidden]="orders?.length > 0">Pas de commandes {{openShippings?'ouvertes':'fermées'}} pour ce mois-ci</h1>
 
   <ion-list>
     
-    <ion-item-sliding *ngFor="let order of monthOrders.get(currentShippingDate?.getTime())?.sort(sortOrdersByCP)  ">
+    <ion-item-sliding *ngFor="let order of orders?.sort(sortOrdersByCP)  ">
 
       <ion-item>
         <ion-grid>
