+++ src/pages/shopper/shopper.ts
@@ -1,11 +1,9 @@
 import { Component, ViewChild } from '@angular/core';
-import { Content, IonicPage, NavController, NavParams } from 'ionic-angular';
+import { Content, IonicPage, ModalController,  } from 'ionic-angular';
 import { LoaderService, Order, OrderService, EnumFinancialStatus, User, UserService } from 'kng2-core';
 import { ShopperItemComponent } from '../../components/shopper-item/shopper-item';
 import { LogisticHeaderComponent }  from '../../components/logistic-header/logistic-header';
-import 'rxjs/Rx';
-import { Observable } from 'rxjs/Rx';
-import { TrackerPage } from '../tracker/tracker';
+import { TrackerPage } from '../../pages/tracker/tracker';
 /**
  * Generated class for the ShopperPage page.
  *
@@ -20,82 +18,19 @@ import { TrackerPage } from '../tracker/tracker';
 export class ShopperPage {
 
   @ViewChild(LogisticHeaderComponent) header: LogisticHeaderComponent;
-
-  isReady: boolean;
-  user: User = new User();
-  nbCabas;
-
-
-  FLOATING = { payment: 'authorized' };  //not yet handled by producers
-  LOCKED = { fulfillments: 'fulfilled,partial' };  //got by the producers (sub-group of FLOATING)
-
-  filtersOrder: any;
-
-  selectedDate: string = new Date().toISOString();
-  currentShippingDate: Date;
-  availableOrders: Date[] = [];
-  monthOrders: Map<number, Order[]> = new Map();
+  
+  private orders: Order[] = [];
 
   constructor(
     private loaderSrv: LoaderService,
     private modalCtrl: ModalController,
-    public navCtrl: NavController,
-    public navParams: NavParams,
     private orderSrv: OrderService,
-    private userSrv: UserService
   ) {
-    this.currentShippingDate = Order.currentShippingDay();
-    this.currentShippingDate.setHours(0, 0, 0);
-    this.filtersOrder = this.FLOATING;
-
   }
 
 
-
-  ngOnInit() {
-    this.loaderSrv.ready().subscribe((loader) => {
-      Object.assign(this.user, loader[1]);
-      this.isReady = true;
-      this.findAllOrdersForShipping();
-    });
-  }
-
-
-  toggleFilter() {
-
-    if (this.filtersOrder.payment) {
-      this.filtersOrder = this.LOCKED;
-    } else {
-      this.filtersOrder = this.FLOATING;
-    }
-    this.findAllOrdersForShipping();
-  }
-
-
-  findAllOrdersForShipping() {
-    let params = { month: (new Date(this.selectedDate).getMonth()) + 1, year: new Date(this.selectedDate).getFullYear() };
-    Object.assign(params, this.filtersOrder);
-    this.monthOrders = new Map();
-    this.availableOrders = [];
-    this.orderSrv.findAllOrders(params).subscribe(orders => {
-      orders.forEach((order: Order) => {
-        order.shipping.when = new Date(order.shipping.when);
-        order.shipping.when.setHours(0, 0, 0)
-        // Object.keys(this.monthOrders)
-        if (!this.monthOrders.has(order.shipping.when.getTime())) {
-          this.monthOrders.set(order.shipping.when.getTime(), []);
-          this.availableOrders.push(order.shipping.when);
-        }
-        this.monthOrders.get(order.shipping.when.getTime()).push(order);
-      });
-      //set currentshipping with first key
-      this.currentShippingDate = new Date(this.monthOrders.keys().next().value);
-
-    })
-  }
-
-  openTracker() {
-    this.modalCtrl.create(TrackerPage, { results: this.monthOrders.get(this.currentShippingDate.getTime()) }).present();
+  displayOrders(orders: Order[]){
+    this.orders = orders;
   }
 
   openTracker4One(order){
